refactor: 프로필 uistate 패키지 추가
